Struct clipmdplus::clip_util::ClipMeta

pub struct ClipMeta {
    pub id_key: u32,
    pub id_system: u32,
    pub data_creation_time: Option<SystemTime>,
    pub data_processed: bool,
    pub data_synced: Option<SystemTime>,
    pub sequence_number: u32,
    pub data_type: String,
    pub clip_data: Option<String>,
    pub clip_i_data: Option<Vec<u8>>,
}

This is the meta data associated with a clipboard clip. This is serialized and points to the IDataObject (clip). The IDataObject may/will be stored independently as a Blob (or bytes).

Fields
id_key: u32

Unique auto-generated key (compliant) (This would have a getter/setter and not be public)

id_system: u32

System Unique ID GUID and MACs are all possible options it was previously not implemented. (Also, this would have a getter/setter and not be public)

data_creation_time: Option<SystemTime>

System time as UTC

data_processed: bool

Updated

data_synced: Option<SystemTime>

Synchronization

sequence_number: u32

Order of clip history

data_type: String

Clipboard data type

clip_data: Option<String>

Internal (this) clipboard data This is an object. In rust an : Arc<Box>,

clip_i_data: Option<Vec<u8>>

Clipboard data object this is the OS object of Format Type in windows a Forms.DataObject

Implementations

impl ClipMeta

fn name() -> &'static str

impl ClipMeta

pub fn new(t: ClipMeta) -> Self

New: Default and T