Module clipmdplus::stdmd::app_error

a learning task. todo Over defined errors are all over the place.

Structs

DbErrorData

DbErrorData. This it temporay. It evolved to this mess from a logging experiment.

DbErrorDataClient

DbErrorDataClient

DbErrorDataDe

DbErrorDataDe

DbErrorDataIo

DbErrorDataIo

DbErrorDataPostgres

DbErrorDataPostgres

DbErrorDataSer

DbErrorDataSer

DbErrorDataStd

DbErrorDataStd

JsonErrorData

JsonErrorData

PhantomError

PhantomData for Errors.

Enums

AppError

App error is a simplified error type. It might be returned to the App. Under review. Might be discarded.

DbError

A custom database error.

GuiError

Errors received from GUI components or upon validation.

JsonError

A custom JsonError needed later.