Module clipmdplus::stdmd::types

types handling library.

Re-exports

pub use self::type_form::*;

pub use self::type_format::*;

pub use self::range::*;

pub use self::sequence::*;

Modules

range

A range object. (todo student)

sequence

A sequence object (todo student)

type_form

Types of forms and UI types.

type_format

The knows format types of entities.

Traits

NamedAny 🔒

Functions

type_is_equal

Compare the types of two Sized objects.

type_name_from_id

Given a TypeId it returns its name as a static str.

type_name_of

The type name done another way. (todo student)

type_name_of_val

The type name of T as a static str.

type_of_val

Returns the TypeId of T.